+++ packages/components/src/components/docs/bal-doc-app/bal-doc-app.tsx
@@ -1,34 +1,40 @@
-import { Component, Host, h, Event, EventEmitter } from '@stencil/core'
-import globalScript from '../../../global'
-import { isBrowser } from '../../../utils/browser'
+import { Component, Host, h, Prop } from '@stencil/core'
+import { isWindowDefined } from '../../../utils/browser'
+import { BalLogger } from '../../../utils/log'
 
 @Component({
   tag: 'bal-doc-app',
   styleUrl: '../../../styles/global.sass',
 })
 export class DocApp {
-  @Event({ bubbles: true, composed: true }) balAppLoad!: EventEmitter<boolean>
+  @Prop() logComponents = ''
+  @Prop() logLifecycle = false
+  @Prop() logEvents = false
+  @Prop() logRender = false
+  @Prop() logCustom = false
 
-  connectedCallback() {
-    globalScript()
-  }
-
-  componentDidLoad() {
-    this.balAppLoad.emit(true)
+  componentDidRender() {
+    const logConfig: BalLogger = {
+      components: this.logComponents
+        .split(',')
+        .map(c => c.trim())
+        .filter(c => c !== ''),
+      lifecycle: this.logLifecycle,
+      event: this.logEvents,
+      render: this.logRender,
+      custom: this.logCustom,
+    }
+    if (isWindowDefined()) {
+      ;(window as any).BaloiseDesignSystem.config.logger = logConfig
+    }
   }
 
   render() {
     return (
-      <Host role="application">
-        <main
-          class={{
-            'bal-app': true,
-            'bal-app--safari': isBrowser('Safari'),
-            'bal-app--touch': isBrowser('touch'),
-          }}
-        >
+      <Host>
+        <bal-app>
           <slot></slot>
-        </main>
+        </bal-app>
       </Host>
     )
   }